Check If Card May Be Removed
Check all top cards to see if there are two or more of the same suit showing. If there are, a card of that suit may be removed. We still need to verify whether the card being removed needs to be lower than another card of the same suit on the board.